Factor out time shift logic from calc.py
This attempts to do so in a way that preserves the automated behavior (encapsulating it completely in the GFDLDataLoader class), but prevents it from happening where it should not by default (i.e. in all other DataLoaders).  One can still request a time offset, but that needs to come from an argument to a Calc object.