MAINT Rename OneDirDataLoader to NestedDictDataLoader and simplify tests
Currently the DictDataLoader and NestedDictDataLoader take absolute filepaths only.  In the future it would not be too hard to add an argument of a "common_root_path," but for now I'm keeping things simple. The renaming is meant to be more descriptive of how each DataLoader actually behaves.  The NestedDictDataLoader is most like the 'one_dir' strategy from the past; however it no longer allows for the input of a dictionary mapping just variable names to file lists -- one needs also to provide an intvl_in.

The DataLoader in the test/data/objects/examples.py has been switched to a DictDataLoader for simplicity and to illustrate the new ability of specifying filesets using wildcards (now that we use xr.open_mfdataset to load files).